get topic counts from s3 other not the database

+++ common/src/main/scala/models/TopicCount.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.{Format, Json}
 
-case class TopicCount(topicName: String,  registrationCount: Long)
+case class TopicCount(topicName: String,  registrationCount: Int)
 
 object TopicCount{
   implicit val topicCountJF: Format[TopicCount] = Json.format[TopicCount]
